backend/pyproject.toml
[project]
name = "nidavellir"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "python-dotenv>=1.0",
    "pydantic-settings>=2.7",
    "pyyaml>=6",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "qdrant-client>=1.9",
]

[tool.uv]
dev-dependencies = [
    "httpx>=0.27",
    "pytest>=8",
    "pytest-asyncio>=0.24",
    "ruff>=0.8",
]
